# [REFACTORY] simplify appender commit