--- a/tsdb/head_bench_test.go
+++ b/tsdb/head_bench_test.go
@@ -27,6 +27,7 @@ import (
 	"github.com/prometheus/prometheus/model/exemplar"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/storage"
+	"github.com/prometheus/prometheus/tsdb/chunkenc"
 	"github.com/prometheus/prometheus/tsdb/chunks"
 	"github.com/prometheus/prometheus/tsdb/wlog"
 )
@@ -88,8 +89,7 @@ func BenchmarkHeadStripeSeriesCreate_PreCreationFailure(b *testing.B) {
 func BenchmarkHead_WalCommit(b *testing.B) {
 	seriesCounts := []int{100, 1000, 10000}
 	series := genSeries(10000, 10, 0, 0)
-	histograms := genHistogramSeries(10000, 10, 0, 119, 1, false)
-	floatHistograms := genHistogramSeries(10000, 10, 0, 119, 1, true)
+	histograms := genHistogramSeries(10000, 11, 0, 119, 1, false)
 
 	for _, seriesCount := range seriesCounts {
 		b.Run(fmt.Sprintf("%d series", seriesCount), func(b *testing.B) {
@@ -112,30 +112,14 @@ func BenchmarkHead_WalCommit(b *testing.B) {
 							}
 						}
 
-						for _, s := range floatHistograms[:seriesCount] {
-							var ref storage.SeriesRef
-							for sampleIndex := int64(0); sampleIndex < samplesPerAppend; sampleIndex++ {
-
-								ref, err = app.AppendHistogram(ref, s.Labels(), ts+sampleIndex, nil, nil)
-								if err != nil {
-									return err
-								}
-
-								_, err = app.AppendExemplar(ref, s.Labels(), exemplar.Exemplar{
-									Labels: labels.FromStrings("trace_id", strconv.Itoa(rand.Int())),
-									Value:  rand.Float64(),
-									Ts:     ts,
-								})
-								if err != nil {
-									return err
-								}
-							}
-						}
-
 						for _, s := range histograms[:seriesCount] {
 							var ref storage.SeriesRef
+							var it chunkenc.Iterator
 							for sampleIndex := int64(0); sampleIndex < samplesPerAppend; sampleIndex++ {
-								ref, err = app.AppendHistogram(ref, s.Labels(), ts+sampleIndex, nil, nil)
+								it = s.Iterator(it)
+								it.Next()
+								_, h := it.AtHistogram(nil)
+								ref, err = app.AppendHistogram(ref, s.Labels(), ts+sampleIndex, h, nil)
 								if err != nil {
 									return err
 								}