@nicolastakashi 2024-10-18
Hey @jesusvazquez @krajorama @ArthurSens,

I just pushed the benchmark tests and would really appreciate your feedback. I’m not entirely confident that I’ve set everything up correctly 😅.

Here’s how I ran the benchmark:

```bash
go test -benchmem -run='^$' -bench='^BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms$' github.com/prometheus/prometheus
/tsdb
```

Below are the results:

**Main Branch:**

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Apple M1 Pro
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/100_series/1_samples_per_append-10         	     958	   1217783 ns/op	   14581 B/op	     253 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/100_series/2_samples_per_append-10         	     594	   2464723 ns/op	   25257 B/op	     454 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/100_series/5_samples_per_append-10         	     286	   5073736 ns/op	   59660 B/op	    1068 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/100_series/100_samples_per_append-10       	      94	  12106552 ns/op	 1128600 B/op	   20618 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/1000_series/1_samples_per_append-10        	     206	   8826639 ns/op	  142690 B/op	    2504 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/1000_series/2_samples_per_append-10        	     172	   7643186 ns/op	  252966 B/op	    4513 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/1000_series/5_samples_per_append-10        	     145	   7919213 ns/op	  592303 B/op	   10590 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/1000_series/100_samples_per_append-10      	      13	  87287128 ns/op	11025601 B/op	  206019 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/10000_series/1_samples_per_append-10       	      44	  24240636 ns/op	 1379661 B/op	   24563 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/10000_series/2_samples_per_append-10       	      37	  34257636 ns/op	 2386759 B/op	   44072 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/10000_series/5_samples_per_append-10       	      19	  62039482 ns/op	 6281212 B/op	  105810 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/10000_series/100_samples_per_append-10     	       2	 761626000 ns/op	110238672 B/op	 2060023 allocs/op
PASS
ok  	github.com/prometheus/prometheus/tsdb	30.034s
```

**PR: 15112**

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Apple M1 Pro
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/100_series/1_samples_per_append-10         	    1029	   1250269 ns/op	   14627 B/op	     253 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/100_series/2_samples_per_append-10         	     672	   2305953 ns/op	   25265 B/op	     455 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/100_series/5_samples_per_append-10         	     297	   5561253 ns/op	   59647 B/op	    1068 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/100_series/100_samples_per_append-10       	     100	  13083397 ns/op	 1127303 B/op	   20618 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/1000_series/1_samples_per_append-10        	     192	   8437737 ns/op	  143375 B/op	    2513 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/1000_series/2_samples_per_append-10        	     186	   8982759 ns/op	  249806 B/op	    4502 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/1000_series/5_samples_per_append-10        	     129	   9172211 ns/op	  595496 B/op	   10615 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/1000_series/100_samples_per_append-10      	      15	  88787250 ns/op	13841601 B/op	  206025 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/10000_series/1_samples_per_append-10       	      58	  22200763 ns/op	 1524710 B/op	   24330 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/10000_series/2_samples_per_append-10       	      38	  37699447 ns/op	 2596134 B/op	   43967 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/10000_series/5_samples_per_append-10       	      19	  67651814 ns/op	 5723811 B/op	  105807 allocs/op
BenchmarkHead_WalCommitWithVaryingSeriesSamplesExemplarsAndHistograms/10000_series/100_samples_per_append-10     	       2	 815671021 ns/op	361808288 B/op	 2060080 allocs/op
PASS
ok  	github.com/prometheus/prometheus/tsdb	30.972s
```

Let me know if you think any adjustments are needed, or if you would like me to rerun the tests with different parameters before we trigger prombench. I’m happy to tweak things as needed!

Looking forward to your feedback!