@nicolastakashi 2024-10-21
Thanks @krajorama and @ArthurSens for your feedback.
I change the code and run benchmark again, this the the benchstat result
```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Apple M1 Pro
                                                      │   main.txt   │                pr.txt                │
                                                      │    sec/op    │    sec/op     vs base                │
Head_WalCommit/100_series/1_samples_per_append-10       1.985m ± ∞ ¹   1.827m ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/100_series/2_samples_per_append-10       3.388m ± ∞ ¹   3.653m ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/100_series/5_samples_per_append-10       8.939m ± ∞ ¹   8.562m ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/100_series/100_samples_per_append-10     19.79m ± ∞ ¹   18.96m ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/1000_series/1_samples_per_append-10      12.53m ± ∞ ¹   12.58m ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/1000_series/2_samples_per_append-10      12.41m ± ∞ ¹   11.36m ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/1000_series/5_samples_per_append-10      16.91m ± ∞ ¹   16.38m ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/1000_series/100_samples_per_append-10    121.6m ± ∞ ¹   119.0m ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/10000_series/1_samples_per_append-10     44.69m ± ∞ ¹   42.45m ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/10000_series/2_samples_per_append-10     51.78m ± ∞ ¹   55.43m ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/10000_series/5_samples_per_append-10     94.62m ± ∞ ¹   86.64m ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/10000_series/100_samples_per_append-10    1.173 ± ∞ ¹    1.127 ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                                 26.05m         25.32m        -2.81%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                                      │   main.txt    │                pr.txt                 │
                                                      │     B/op      │     B/op       vs base                │
Head_WalCommit/100_series/1_samples_per_append-10       27.80Ki ± ∞ ¹   27.85Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/100_series/2_samples_per_append-10       40.25Ki ± ∞ ¹   40.56Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/100_series/5_samples_per_append-10       79.55Ki ± ∞ ¹   79.37Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/100_series/100_samples_per_append-10     1.424Mi ± ∞ ¹   1.417Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/1000_series/1_samples_per_append-10      273.6Ki ± ∞ ¹   272.3Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/1000_series/2_samples_per_append-10      405.9Ki ± ∞ ¹   419.5Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/1000_series/5_samples_per_append-10      848.2Ki ± ∞ ¹   844.7Ki ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/1000_series/100_samples_per_append-10    15.21Mi ± ∞ ¹   15.25Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/10000_series/1_samples_per_append-10     2.668Mi ± ∞ ¹   2.548Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/10000_series/2_samples_per_append-10     3.897Mi ± ∞ ¹   4.249Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/10000_series/5_samples_per_append-10     8.084Mi ± ∞ ¹   9.131Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/10000_series/100_samples_per_append-10   388.2Mi ± ∞ ¹   388.2Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                                 1.146Mi         1.165Mi        +1.60%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                                      │   main.txt   │                pr.txt                │
                                                      │  allocs/op   │  allocs/op    vs base                │
Head_WalCommit/100_series/1_samples_per_append-10        441.0 ± ∞ ¹    442.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/100_series/2_samples_per_append-10        643.0 ± ∞ ¹    644.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/100_series/5_samples_per_append-10       1.268k ± ∞ ¹   1.267k ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/100_series/100_samples_per_append-10     20.86k ± ∞ ¹   20.86k ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/1000_series/1_samples_per_append-10      4.337k ± ∞ ¹   4.327k ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/1000_series/2_samples_per_append-10      6.316k ± ∞ ¹   6.340k ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/1000_series/5_samples_per_append-10      12.43k ± ∞ ¹   12.36k ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/1000_series/100_samples_per_append-10    208.3k ± ∞ ¹   208.4k ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/10000_series/1_samples_per_append-10     41.52k ± ∞ ¹   40.31k ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/10000_series/2_samples_per_append-10     60.05k ± ∞ ¹   61.87k ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/10000_series/5_samples_per_append-10     118.7k ± ∞ ¹   120.9k ± ∞ ¹       ~ (p=1.000 n=1) ²
Head_WalCommit/10000_series/100_samples_per_append-10   2.070M ± ∞ ¹   2.070M ± ∞ ¹       ~ (p=1.000 n=1) ³
geomean                                                 16.20k         16.22k        +0.14%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
³ all samples are equal
```