@krajorama 2024-10-22
My measurement. Please always use `-count 6` or higher when running benchmarks to get more data.
```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/tsdb
cpu: AMD Ryzen 7 4700U with Radeon Graphics         
                                                     │   main.txt   │                 pr.txt                 │
                                                     │    sec/op    │     sec/op      vs base                │
Head_WalCommit/100_series/1_samples_per_append-8       399.3µ ±  2%    977.1µ ±  59%  +144.70% (p=0.026 n=6)
Head_WalCommit/100_series/2_samples_per_append-8       2.175m ± 61%    2.569m ±   5%   +18.15% (p=0.015 n=6)
Head_WalCommit/100_series/5_samples_per_append-8       7.169m ± 15%    7.676m ± 478%         ~ (p=0.240 n=6)
Head_WalCommit/100_series/100_samples_per_append-8     13.41m ±  1%    14.54m ±   3%    +8.47% (p=0.002 n=6)
Head_WalCommit/1000_series/1_samples_per_append-8      10.29m ±  3%    10.32m ±   2%         ~ (p=1.000 n=6)
Head_WalCommit/1000_series/2_samples_per_append-8      10.63m ±  4%    10.44m ±   7%         ~ (p=0.589 n=6)
Head_WalCommit/1000_series/5_samples_per_append-8      11.02m ± 19%    10.79m ± 554%         ~ (p=0.394 n=6)
Head_WalCommit/1000_series/100_samples_per_append-8    132.6m ±  1%    139.3m ±   2%    +5.08% (p=0.002 n=6)
Head_WalCommit/10000_series/1_samples_per_append-8     39.43m ±  1%    46.76m ±   4%   +18.59% (p=0.002 n=6)
Head_WalCommit/10000_series/2_samples_per_append-8     54.32m ±  1%    62.01m ±   9%   +14.15% (p=0.002 n=6)
Head_WalCommit/10000_series/5_samples_per_append-8     95.69m ±  0%   106.63m ±   2%   +11.43% (p=0.002 n=6)
Head_WalCommit/10000_series/100_samples_per_append-8    1.372 ±  2%     1.452 ±   2%    +5.90% (p=0.002 n=6)
geomean                                                19.85m          22.86m          +15.16%

                                                     │    main.txt    │                pr.txt                │
                                                     │      B/op      │      B/op       vs base              │
Head_WalCommit/100_series/1_samples_per_append-8       16.08Ki ±   0%   16.08Ki ±   0%       ~ (p=0.483 n=6)
Head_WalCommit/100_series/2_samples_per_append-8       29.64Ki ±   0%   29.65Ki ±   0%       ~ (p=0.420 n=6)
Head_WalCommit/100_series/5_samples_per_append-8       68.45Ki ±   1%   68.58Ki ±   0%       ~ (p=0.240 n=6)
Head_WalCommit/100_series/100_samples_per_append-8     1.393Mi ±   1%   1.442Mi ±   3%  +3.54% (p=0.002 n=6)
Head_WalCommit/1000_series/1_samples_per_append-8      150.3Ki ±   2%   153.2Ki ±   2%       ~ (p=0.115 n=6)
Head_WalCommit/1000_series/2_samples_per_append-8      288.4Ki ±   1%   291.9Ki ±   3%  +1.22% (p=0.002 n=6)
Head_WalCommit/1000_series/5_samples_per_append-8      730.6Ki ±   1%   738.1Ki ±   3%       ~ (p=0.100 n=6)
Head_WalCommit/1000_series/100_samples_per_append-8    15.06Mi ±   0%   16.31Mi ±   8%       ~ (p=0.589 n=6)
Head_WalCommit/10000_series/1_samples_per_append-8     1.522Mi ±   4%   1.522Mi ±  10%       ~ (p=1.000 n=6)
Head_WalCommit/10000_series/2_samples_per_append-8     2.946Mi ±   0%   2.946Mi ±  14%       ~ (p=0.615 n=6)
Head_WalCommit/10000_series/5_samples_per_append-8     7.259Mi ±   0%   7.259Mi ±   0%       ~ (p=0.894 n=6)
Head_WalCommit/10000_series/100_samples_per_append-8   145.8Mi ± 165%   145.8Mi ± 165%       ~ (p=0.937 n=6)
geomean                                                837.2Ki          848.3Ki         +1.33%

                                                     │  main.txt   │               pr.txt                │
                                                     │  allocs/op  │  allocs/op   vs base                │
Head_WalCommit/100_series/1_samples_per_append-8        318.0 ± 0%    318.0 ± 0%       ~ (p=1.000 n=6) ¹
Head_WalCommit/100_series/2_samples_per_append-8        528.0 ± 0%    528.0 ± 0%       ~ (p=1.000 n=6) ¹
Head_WalCommit/100_series/5_samples_per_append-8       1.158k ± 0%   1.159k ± 0%  +0.09% (p=0.013 n=6)
Head_WalCommit/100_series/100_samples_per_append-8     20.76k ± 0%   20.76k ± 0%       ~ (p=0.058 n=6)
Head_WalCommit/1000_series/1_samples_per_append-8      3.057k ± 1%   3.092k ± 1%       ~ (p=0.481 n=6)
Head_WalCommit/1000_series/2_samples_per_append-8      5.109k ± 0%   5.130k ± 1%       ~ (p=0.182 n=6)
Head_WalCommit/1000_series/5_samples_per_append-8      11.31k ± 0%   11.31k ± 0%       ~ (p=0.727 n=6)
Head_WalCommit/1000_series/100_samples_per_append-8    206.6k ± 0%   206.6k ± 0%       ~ (p=0.182 n=6)
Head_WalCommit/10000_series/1_samples_per_append-8     30.04k ± 0%   30.04k ± 0%       ~ (p=0.455 n=6)
Head_WalCommit/10000_series/2_samples_per_append-8     50.05k ± 0%   50.05k ± 0%       ~ (p=1.000 n=6)
Head_WalCommit/10000_series/5_samples_per_append-8     110.1k ± 0%   110.1k ± 0%       ~ (p=1.000 n=6) ¹
Head_WalCommit/10000_series/100_samples_per_append-8   2.060M ± 0%   2.060M ± 0%       ~ (p=0.857 n=6)
geomean                                                13.86k        13.88k       +0.14%
¹ all samples are equal
```